Set AAR after call is established
Manual tests showed that sometimes on pre Android 11 devices (not on all of them) AAR which uses under the hood [startBluetoothSco](https://developer.android.com/reference/android/media/AudioManager#startBluetoothSco()) method doesn't work when communication mode is NOT set. What we've observed is that although `SCO_CONNECTED` event is issued by the OS the audio still routes to SPEAKER instead of Bluetooth device.

Setting the communication mode is done internally by the SDK once the call is established. 

As a solution delay setting the communication mode in the reference app until `onCallEstablished` is called.