Improve CallKit UI when incoming call arrives during an ongoing call
This MR tries to address iOS specific issue connected with VoIP push requirements on iOS.

By design, whenever a new VoIP push arrives the customer is obliged to forward this push to SDK (`SinchClient`). Then, internally we might automatically deny the call and decide not to invoke `didReceiveIncomingCall` delegate's callback.

This behaviour and design is not a problem on Android since customer app may decide if it want to ignore the notification (and show it only based on `didReceiveIncomingCall`), however on iOS we can't do this as we are obliged to report each incoming VoIP to OS by calling `reportNewIncomingCallWithUUID`. This, then triggers CallKit framework UI showing a new incoming call. But once the push is relayed to SinchClient, the internal logic might deny the call and never trigger `didReceiveIncomingCall`. This leaves UI in wrong state when user sees that there's a new call while in fact it was already denied.