Fix no information about incoming call when screen is locked in Android
This MR fixes the bug in Android reference app that caused no information about incoming call when the screen is locked.

The issue was caused by trying to workaround the recommended flow and show the incoming call screen automatically if the application was in foreground. Now everything is handled by system notification.